feat(cart): 아이템 삭제 service 메소드 구현

+++ cartpost-service/src/main/java/com/example/cartpostservice/cart/service/CartServiceImpl.java
@@ -61,9 +61,13 @@ public class CartServiceImpl implements CartService {
         CartItemsEntity cartItem = cartItemsRepository.findByCode(itemCode)
                 .orElseThrow(() -> new BusinessException(CustomStatusCode.NOT_FOUND_ITEM));
 
+        if (!cart.getCode().equals(cartItem.getCartCode())) {
+            throw new BusinessException(CustomStatusCode.FORBIDDEN_ITEM);
+        }
+
         cartItemsRepository.delete(cartItem);
 
-        return null;
+        return EmptyResponse.getInstance();
     }
 
     private int calculateTotalAmount(Instant startedAt, Instant endedAt, PaymentType paymentType, String amount) {
